Queue is a linear data structure
Queue is a data structure in which the data can be added and removed from 2 seperate ends. Data is added in queue from rear end and data is removed from front end. Operation of addition of data in queue is known as ENQUEUE and deletion of data is known as DEQUEUE.